Updated Digester to allow just setting a SAX2 Reader into Digester.  This completly gets around Weblogic 6.0 locking down the JAXP 1.0 distribution.
All you need to have is SAX2 somewhere in the classpath, and then create an
XMLReader, and set it in.  This will bypass Digester's namespace and
validation settings, but at least it will work.

The offending call, BTW, is the call to SAXParser.parse.  In JAXP 1.0,
it takes a DefaultHandler as an argument, and was updated in JAXP 1.1.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/digester/trunk@132279 13f79535-47bb-0310-9956-ffa450edef68