Add support for public getFeature() and setFeature() methods to configure the SAXParserFactory.  Also added getProperty() and setProperty() methods to do the same sort of thing for SAXParser properties.  (I changed the original suggested patch to return SAXNotRecognizedException and SAXNotSupportedException errors to the caller).
The current implementation still tries to set the schema location and source
properties, even on a JAXP/1.1 parser that does not recognize them.  Shouldn't
this be the responsibility of the caller?  Note that if they are removed, we
should definitely add an example to the o.a.c.digester package description
because this will likely be a common use case in a JAXP/1.2 world.

Submitted by:	Jean-francois Arcand <jeanfrancois.arcand at sun.com>


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/digester/trunk@132376 13f79535-47bb-0310-9956-ffa450edef68