XMLSignatureInput and Reference now support a debuggingFunctionality:
XMLSignatureInput offers the methods

  String getHTMLRepresentation
  String getHTMLRepresentation(Set inclusiveNamespaces)

which do work if the internal representation is a node set. They return an HTML representation of the node set.



When a Reference contains Transforms, the new getNodesetBeforeFirstCanonicalization returns the last existing node set. This allows to debug canonicalization algorithms more easily and allows users to check which node set has been canonicalized.


git-svn-id: https://svn.apache.org/repos/asf/xml/security/trunk@350556 13f79535-47bb-0310-9956-ffa450edef68