As agreed to earlier on the COMMONS-DEV mailing list, remove direct dependence on commons-collections.  This involves maintaining a local copy of ArrayStack that was originally contributed to Collections when Digester and Beanutils were factored out of Struts.
There is still a runtime dependency on commons-collections by virtue of
the fact that commons-beanutils requires it.  Once this dependency is
lifted, we can remove commons-collections settings from the build.xml and
build.properties.sample files, as well as from the unit test classpath.

Also fixed a couple of Javadoc errors that triggered warning messages.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/digester/trunk@132554 13f79535-47bb-0310-9956-ffa450edef68