- Override hashCode when overriding equals - Adding simple testcases
Some tests are ignored because I'm not sure if the equals method really does what is wanted:
A Collection is compared by collection1.equals(collection2) which is "false" in almost every case as a reference comparision is done [1].
I'm not sure if this is really wanted. If so, the tests annotated with @ignore can be removed. Otherwise a fix for the related equals method is needed.

[1] http://java.sun.com/j2se/1.5.0/docs/api/java/util/Collection.html#equals%28java.lang.Object%29

git-svn-id: https://svn.apache.org/repos/asf/directory/shared/trunk@946658 13f79535-47bb-0310-9956-ffa450edef68