Restored jdk 1.3 compatibility to StringUtils
This implementation is probably also faster than the non 1.3 compat one.

Please note that I do not make any claims as to keeping 1.3 compatibility
for the rest of plexus-utils. But this class is used by everyone and
only contained a few lines of 1.4 code

Improved test coverage on method in question