Fix Issue #235: Deleting flows doesn't delete all the flow entries
Add missing setting for the OpenFlow field out_port in
the "Modify Flow Entry Message". Appararently, its value
is important for DELETE or DELETE_STRICT messages if
in the matching includes checking of the output port.

If the matching does NOT include the output port,
its value must be set to OFPP_NONE (0xffff for OpenFlow 1.0).
Previously, it was initialized to zero (a valid port number),
and the deleting of flow entries would succeed only if the outgoing
port number was zero.