Fixes toward Bug #310 and Bug #318 (and might be Bug #328).
FlowManager refactoring to simplify and speedup the code and
eliminate Titan DB access bugs.

Removed the heuristic algorithm to automatically delete the Flow object
from the Network MAP, because it wasn't working with multiple
controllers.
This removal probably also fixes a bug (not identifed yet?) when a
flow monitored by ONOS could actually disappear from the Network MAP
when the Shortest Path fails.