Fixes toward Bug #310 and Bug #318 (and might be Bug #328).
FlowManager refactoring to simplify and speedup the code and
eliminate Titan DB access bugs.

Add new method reconcileFlow() that reconciles a single flow
properly by avoiding a number of bugs in the older
implementation in method reconcileFlows().
One of the fixes is avoiding inappropriate closing of Titan DB
transaction.

With Masa's CHO tests for link failure on DEV-Z cluster,
there was only a single allping failure in almost 4 hours.
The root cause for that ping failure didn't seem to be flow-related.
With the older implementation, there was allping failure
every 10-15 minutes.