Work toward fixing Bug ONOS-859:
If there is a Titan exception when adding or deleting Flows to the GraphDB,
keep retrying until success.

NOTE: Most likely this issue is NOT the root cause for (most of)
the missing flows as reported by Bug ONOS-859.
Initial investigation on 4-node VM setup in the lab couldn't reproduce
the problem. We need to investigate the root cause on the 3-node
Baremetal cluster after it becomes available.

Reviewed by: Pankaj