Workaround for Outgoing-Link may be out-dated problem
When Link is removed as side-effect of Port remove,
check if the ONOS instance owns the Dst DPID, and remove Link from DB,
only if the instance is in control.

This workaround only makes sense if
ZK check hasControl() has lower latency then
removing a Link, which does not exist on DataStore.

Change-Id: I23586f09e925ceb817031fcef066b9a98abb84b8