Cherry-pick from https://gerrit.onos.onlab.us/#/c/92/4 (partially).
NOTE: Abandoned the implementation about automatic-barrier function.

FlowEntryAdded event is changed to be occurred after reception of BARRIER_REPLY (ONOS-915).
Fixed a bug that the same FlowEntry is installed to switch multiple times.

Change-Id: Ibb5cdaf9a478e2697232203d9190dedb4792f108

Fixed a bug that FlowEntries are shallow-copied in FlowPusher.

Change-Id: I704d245d3660a91f5e60fd6cdf05f89b1fba6c25

Fixed a problem that FlowPusherTest fails often(ONOS-984).

Change-Id: I2a290d5dc756e43a8058c6f35372f2e2932a8d73