Fix setRecompressAddedZips has no effect on the resulting archive
To check if a given entry is zip file or not the first four bit
should be read. There is an issue when entries are being added
asynchronously. Because it's faster to submit entries
than to compress them, input streams are opened faster than
they are closed. Eventually this could lead to too many
open files error.

There was a workaround for this issue. Wrapper around the
`InputStreamSupplier` so the first four bytes are read
when the entry is compressed and not when it's submitted.
That solves the too many open files problem, but
unfortunately has no effect on the resulting archive.
The compression method is determined when the entry is
submitted so changing it afterwards has no effect.

Use the newly introduced `ZipArchiveEntryRequestSupplier`
to supply the whole `ZipArchiveEntryRequest`
(including the compression method) when the entry is
compressed.

This fixes #53 and closes #55