Fix symbolic links not properly encoded in ZIP archives
To encode the symbolic links `AbstractZipArchiver` uses
`ZipEncoding` but it does not use it properly.
It uses the whole backing array of the returned ByteBuffer
while it it should use it only up to the buffer limit.
This may lead to additional characters at the end of the
symbolic link.

This does not manifest as a bug because Common Compress up to 1.14
returns a ByteBuffer that wraps a byte array containing the
encoded payload. But Common Compress 1.15 returns new
implementation and we need to fix it before upgrading.
Actually that is how it was found.

This closes #73