Upgrade hibernate to version 5.4
The hibernate implementation of
javax.persistence.criteria.CriteriaBuilder.In in version 5.3 generates
syntactically invalid SQL (IN() without parameters) for PostgreSQL
if the value() method is given an empty list. This has been fixed
in hibernate-orm rev. ab9ae431858dc6727023d7f03bd6925f99011c62 and
thus in version 5.4.10 and above.

LADA server was affected at least in the TagService, where a predicate
is build with userInfo.getMessstellen(), which is possibly empty.

Removed some outdated comments in Dockerfile, since we are using JTS
1.16.0 from Debian buster.