Input validation for LAF import service
Check for valid encoding name and base64 file content.

Object.toString() did not actually return valid base64 for the used
implementation of JsonValue, but returned the value enclosed in double
quotes. Although the commons.codec decoder did not complain and managed
to decode the string, somehow, nevertheless use javax.json API to
retrieve the string correctly and replace the (unstated, with regard to
pom.xml) dependency to commons.codec with Java's own decoder.