Add extra database and data sources for automated tests
This allows running automated tests and working with test data in the
normal deployment in parallel with a single installation. Because the
automated tests need an empty database, before it was necessary to e.g.
stop Wildfly, drop and recreate the database and restart Wildfly to
switch between running automated tests and working with the extensive
test data set.